move helper functions to helpers.go
